requirements.txt
absl-py==2.0.0
arch==5.6.0
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.3
brotlipy==0.7.0
cachetools==5.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.7
cycler==0.11.0
dask==2023.5.0
decorator==5.1.1
distributed==2023.5.0
einops==0.7.0
executing==2.0.1
filelock==3.12.4
finch-clust==0.2.0
flatbuffers==23.5.26
fonttools==4.42.1
frozendict==2.3.10
fsspec==2023.9.1
ftfy==6.1.3
gast==0.4.0
gdown==5.1.0
google-auth==2.25.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.17.1
hurst==0.0.5
idx2numpy==1.2.3
importlib-metadata==6.8.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
libclang==16.0.6
lit==17.0.4
llvmlite==0.41.1
locket==1.0.0
loralib==0.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-service==2.4.0
more-itertools==10.1.0
msgpack==1.0.7
natsort==8.4.0
numba==0.58.1
numpy==1.24.3 
oauthlib==3.2.2
openai-whisper==20230918
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
partd==1.4.1
patsy==0.5.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.43
property-cached==1.6.4
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.1
semver==3.0.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
stumpy==1.12.0
tblib==3.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
TFSnippet @ git+https://github.com/haowen-xu/tfsnippet.git@63adaf04d2ffff8dec299623627d55d4bacac598
threadpoolctl==3.2.0
tiktoken==0.3.3
timm==0.9.7
tokenizers==0.14.1
tomli==2.0.1
toolz==0.12.0
# torch==1.12.1
# torchvision==0.13.1
torchviz==0.0.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.34.0
triton==2.0.0
tsfresh==0.20.1
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
zhusuan @ git+https://github.com/thu-ml/zhusuan.git@4386b2a12ae4f4ed8e694e504e51d7dcdfd6f22a
zict==3.0.0
zipp==3.17.0




python main.py \
  --dataset iDigits \
  --data-path /local_datasets \
  --num_tasks 10 \
  --batch-size 24 \
  --epochs 5 \
  --protocol joint_block \
  --jb-seed 123 \
  --seed 42 \
  --output_dir /root/VIL_main12/output1/jb_iDigits

python main_all.py --dataset DomainNet --batch-size 24 --versatile_inc --random_inc --num_tasks 345 --n_tasks 15 --shuffle True --seed 123 --output_dir /root/VIL_main12/output1/output_ril_Do  --IC --thre 0.0 --beta 0.01 --use_cast_loss --k 2 --d_threshold --use_replay --replay_size 10000 --replay_ratio 0.30   --rebalance   --rb_level class_domain   --rb_mode effective_num   --rb_beta 0.9999   --rb_cap 5   --rb_warmup_epochs 3  --replay_drift --drift_weight 0.6 --drift_alpha 0.7   --drift_hist_len 8 --drift_min_count 32 --drift_temperature 0.8 --drift_anchor_weight 0.9

